+++ crates/lsp/src/lib.rs
@@ -313,6 +313,73 @@ fn on_code_lens_request(
                 lenses.push(compile_lens);
             }
         }
+
+        if package.is_binary() {
+            if let Some(main_func_id) = context.get_main_function(&crate_id) {
+                let location = context.function_meta(&main_func_id).name.location;
+                let file_id = location.file;
+
+                // Ignore diagnostics for any file that wasn't the file we saved
+                // TODO: In the future, we could create "related" diagnostics for these files
+                // TODO: This currently just appends the `.nr` file extension that we store as a constant,
+                // but that won't work if we accept other extensions
+                if fm.path(file_id).with_extension(FILE_EXTENSION) != file_path {
+                    continue;
+                }
+
+                let range = byte_span_to_range(files, file_id.as_usize(), location.span.into())
+                    .unwrap_or_default();
+
+                let command = Command {
+                    title: format!("{ARROW} {COMPILE_CODELENS_TITLE}"),
+                    command: COMPILE_COMMAND.into(),
+                    arguments: Some(vec![
+                        "--program-dir".into(),
+                        format!("{}", workspace.root_dir.display()).into(),
+                        "--package".into(),
+                        format!("{}", package.name).into(),
+                    ]),
+                };
+
+                let lens = CodeLens { range, command: command.into(), data: None };
+
+                lenses.push(lens);
+            }
+        }
+
+        if package.is_contract() {
+            // Currently not looking to deduplicate this since we don't have a clear decision on if the Contract stuff is staying
+            for contract in context.get_all_contracts(&crate_id) {
+                let location = contract.location;
+                let file_id = location.file;
+
+                // Ignore diagnostics for any file that wasn't the file we saved
+                // TODO: In the future, we could create "related" diagnostics for these files
+                // TODO: This currently just appends the `.nr` file extension that we store as a constant,
+                // but that won't work if we accept other extensions
+                if fm.path(file_id).with_extension(FILE_EXTENSION) != file_path {
+                    continue;
+                }
+
+                let range = byte_span_to_range(files, file_id.as_usize(), location.span.into())
+                    .unwrap_or_default();
+
+                let command = Command {
+                    title: format!("{ARROW} {COMPILE_CODELENS_TITLE}"),
+                    command: COMPILE_COMMAND.into(),
+                    arguments: Some(vec![
+                        "--program-dir".into(),
+                        format!("{}", workspace.root_dir.display()).into(),
+                        "--package".into(),
+                        format!("{}", package.name).into(),
+                    ]),
+                };
+
+                let lens = CodeLens { range, command: command.into(), data: None };
+
+                lenses.push(lens);
+            }
+        }
     }
 
     let res = if lenses.is_empty() { Ok(None) } else { Ok(Some(lenses)) };
@@ -419,7 +486,7 @@ fn on_did_save_text_document(
             let fm = &context.file_manager;
             let files = fm.as_simple_files();
 
-            for FileDiagnostic { file_id, diagnostic } in file_diagnostics {
+            for FileDiagnostic { file_id, diagnostic, call_stack: _ } in file_diagnostics {
                 // Ignore diagnostics for any file that wasn't the file we saved
                 // TODO: In the future, we could create "related" diagnostics for these files
                 // TODO: This currently just appends the `.nr` file extension that we store as a constant,